+++ src/translations/en/embeddedassets.php
@@ -13,4 +13,6 @@ return [
     'Video' => 'Video',
     'Use YouTube nocookie?' => 'Use YouTube nocookie?',
     'Whether to force YouTube video iframes to use the `youtube-nocookie.com` domain.' => 'Whether to force YouTube video iframes to use the `youtube-nocookie.com` domain.',
+    'Disable tracking on Vimeo iframes?' => 'Disable tracking on Vimeo iframes?',
+    'Whether to force Vimeo video iframes to use the query parameter `dnt=1` to prevent tracking.' => 'Whether to force Vimeo video iframes to use the query parameter `dnt=1` to prevent tracking.',
 ];